// src/app/models/clima.model.ts
export class Clima {
    lugar: string;
    temperatura: number;
    min: number;
    max: number;
    humedad: number;
    codigo: string;
    guardado: boolean;
    descripcion: string;

    // constructor(lugar, temperatura, min, max, humedad) {
    //     this.lugar = lugar;
    //     this.temperatura = temperatura;
    //     this.min = min;
    //     this.max = max;
    //     this.humedad = humedad;
    // }

    constructor() {
        
    }

}